Async execution in a separate thread. (#434)
* tk_async_execute

* Refactoring
