+++ requirements/mandatory.txt
@@ -2,3 +2,4 @@ aiohttp>=3.6.0,<3.9.0
 aiohttp_socks==0.8.0
 typeguard==2.13
 typing_extensions>=4,<5; python_version < "3.11"
+tkinter-async-execute==1.0.0
