+++ src/daf_gui/extra.py
@@ -19,6 +19,8 @@ from ttkbootstrap.tooltip import ToolTip
 
 import datetime as dt
 
+import tk_async_execute as tae
+
 
 class AdditionalWidget:
     """
@@ -117,7 +119,7 @@ def setup_additional_live_update(w: ttk.Button, frame):
             connector = get_connection()
             await connector.execute_method(old.real_object, "update", **values)
 
-        async_execute(update(), parent_window=frame.origin_window)
+        tae.async_execute(update(), wait=False, pop_up=True, master=frame.origin_window)
 
     w.configure(command=_callback)
     ToolTip(w, "Update the actual object with new parameters (taken from this window)", topmost=True)
@@ -146,7 +148,7 @@ def setup_additional_live_refresh(w: ttk.Button, frame):
 
                 frame_.remember_gui_data()
 
-        async_execute(refresh(), parent_window=frame.origin_window)
+        tae.async_execute(refresh(), wait=False, pop_up=True, master=frame.origin_window)
 
     w.configure(command=_callback)
     ToolTip(w, "Load updated values from the object into the window", topmost=True)
